changed counter value to int64

+++ measurements/counters.go
@@ -4,13 +4,13 @@ package measurements
 type Counter struct {
 	Prefix string
 	Name   string
-	Value  int
+	Value  int64
 	Rate   float32
 	Tags   map[string]string
 }
 
 // NewCounter returns a pointer to Counter that can be sent to a serializer.
-func NewCounter(prefix, name string, value int, rate float32, tags map[string]string) *Counter {
+func NewCounter(prefix, name string, value int64, rate float32, tags map[string]string) *Counter {
 	return &Counter{
 		Prefix: prefix,
 		Name:   name,
